# src/main/resources/config-openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Reservation Microservice API
  description: Microservicio para gestión de reservas y servicios asociados a las reservas.
  version: 1.0.0

servers:
  - url: http://localhost:8081/api/v1
    description: Servicio de Gestión de Reservas

paths:
  # ==================== RESERVATIONS ====================
  /reservations:
    get:
      tags:
        - Reservation
      summary: Obtener todas las reservas con paginación y filtros
      operationId: getReservations
      parameters:
        - name: page
          in: query
          required: false
          description: Número de página (base 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: Tamaño de la página
          schema:
            type: integer
            default: 20
        - name: ownerId
          in: query
          required: false
          description: Filtrar por ID del dueño
          schema:
            type: integer
            format: int64
        - name: carerId
          in: query
          required: false
          description: Filtrar por ID del cuidador
          schema:
            type: integer
            format: int64
        - name: state
          in: query
          required: false
          description: Filtrar por estado de la reserva (PENDING, CONFIRMED, CANCELLED)
          schema:
            type: string
        - name: fromDate
          in: query
          required: false
          description: Filtrar por fecha mínima del servicio
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          required: false
          description: Filtrar por fecha máxima del servicio
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Reservas obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponseList'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    post:
      tags:
        - Reservation
      summary: Crear una nueva reserva
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequestDTO'
      responses:
        '201':
          description: Reserva creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponseDTO'
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor

  /reservations/{id}:
    get:
      tags:
        - Reservation
      summary: Obtener una reserva por ID
      operationId: getReservationById
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la reserva
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Reserva obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponseDTO'
        '404':
          description: Reserva no encontrada
        '500':
          description: Error interno del servidor

    put:
      tags:
        - Reservation
      summary: Actualizar una reserva existente
      operationId: updateReservation
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la reserva
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequestDTO'
      responses:
        '200':
          description: Reserva actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponseDTO'
        '404':
          description: Reserva no encontrada
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor

    delete:
      tags:
        - Reservation
      summary: Eliminar lógicamente una reserva
      operationId: deleteReservation
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la reserva
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Reserva eliminada exitosamente
        '404':
          description: Reserva no encontrada
        '500':
          description: Error interno del servidor


  # ==================== RESERVATION SERVICES ====================
  /reservation-services:
    get:
      tags:
        - ReservationService
      summary: Obtener los servicios asociados a las reservas con paginación
      operationId: getReservationServices
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: reservationId
          in: query
          required: false
          description: Filtrar por ID de la reserva
          schema:
            type: integer
            format: int64
        - name: serviceId
          in: query
          required: false
          description: Filtrar por ID del servicio
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Servicios de reserva obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationServiceResponseList'
        '500':
          description: Error interno del servidor

    post:
      tags:
        - ReservationService
      summary: Crear un nuevo vínculo entre reserva y servicio
      operationId: createReservationService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationServiceRequestDTO'
      responses:
        '201':
          description: Servicio de reserva creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationServiceResponseDTO'
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor

  /reservation-services/{id}:
    get:
      tags:
        - ReservationService
      summary: Obtener un servicio de reserva por ID
      operationId: getReservationServiceById
      parameters:
        - name: id
          in: path
          required: true
          description: ID del servicio de reserva
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Servicio de reserva obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationServiceResponseDTO'
        '404':
          description: Servicio de reserva no encontrado
        '500':
          description: Error interno del servidor

    delete:
      tags:
        - ReservationService
      summary: Eliminar lógicamente un servicio de reserva
      operationId: deleteReservationService
      parameters:
        - name: id
          in: path
          required: true
          description: ID del servicio de reserva
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Servicio de reserva eliminado exitosamente
        '404':
          description: Servicio de reserva no encontrado
        '500':
          description: Error interno del servidor


components:
  schemas:
    # ==================== COMMON ====================
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    BaseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64

    AbstractDTO:
      allOf:
        - $ref: '#/components/schemas/BaseDTO'
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            active:
              type: boolean

    # ==================== RESERVATION ====================
    ReservationRequestDTO:
      type: object
      required:
        - ownerId
        - carerId
        - serviceDate
        - state
      properties:
        ownerId:
          type: integer
          format: int64
        carerId:
          type: integer
          format: int64
        serviceDate:
          type: string
          format: date-time
        state:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
        reservationServiceIds:
          type: array
          items:
            type: integer
            format: int64

    ReservationResponseDTO:
      allOf:
        - $ref: '#/components/schemas/AbstractDTO'
        - type: object
          properties:
            ownerId:
              type: integer
              format: int64
            carerId:
              type: integer
              format: int64
            serviceDate:
              type: string
              format: date-time
            state:
              type: string
            reservationServices:
              type: array
              items:
                $ref: '#/components/schemas/ReservationServiceResponseDTO'

    ReservationResponseList:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ReservationResponseDTO'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    # ==================== RESERVATION SERVICE ====================
    ReservationServiceRequestDTO:
      type: object
      required:
        - reservationId
        - serviceId
      properties:
        reservationId:
          type: integer
          format: int64
        serviceId:
          type: integer
          format: int64

    ReservationServiceResponseDTO:
      allOf:
        - $ref: '#/components/schemas/AbstractDTO'
        - type: object
          properties:
            reservationId:
              type: integer
              format: int64
            serviceId:
              type: integer
              format: int64

    ReservationServiceResponseList:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ReservationServiceResponseDTO'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer